UC7 Issue #26 progress
made the users name clickable to delete a user, backend does not have a
api endpoint to edit yet, when that happens i will create a new
component to update / delete a user like for items. Currently the delete
does not work, something to do with the database not allowing it.